requirements.txt
# Core parsing & utilities
beautifulsoup4
lxml
urlextract
python-dotenv
requests
pyyaml
rich
